refactor: move namespace Core to Domain framework (#55)
