--- Projects/Core/Sources/Namespace/ChoiceNamespace.swift
@@ -1,16 +1,0 @@
-//
-//  ChoiceNamespace.swift
-//  Core
-//
-//  Created by 박소윤 on 2023/12/11.
-//  Copyright © 2023 AB. All rights reserved.
-//
-
-import Foundation
-
-public enum ChoiceTemp {
-    public enum Option: String {
-        case A
-        case B
-    }
-}
